requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
black==22.1.0
blue==0.9.1
bytecode==0.17.0
certifi==2025.8.3
cffi==1.17.1
click==8.2.1
coverage==7.10.7
cryptography==45.0.6
ddtrace==4.1.3
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
envier==0.6.1
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
flake8==4.0.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.1
iniconfig==2.1.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mccabe==0.6.1
mdurl==0.1.2
mypy_extensions==1.1.0
opentelemetry-api==1.39.1
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pyasn1==0.6.1
pycodestyle==2.8.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyflakes==2.4.0
Pygments==2.19.2
PyMySQL==1.1.1
pytest==8.4.2
pytest-cov==7.0.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
rsa==4.9.1
sentry-sdk==2.34.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.2
tomli==2.2.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==2.0.1
zipp==3.23.0
